+++ lib/src/util/embedjs.dart
@@ -7,6 +7,7 @@ import 'dart:async';
 import 'dart:html';
 import 'dart:js';
 
+
 /**
  * Embeds a JavaScript source file into the current document.  This does
  * not wait for the script to load and be parsed.
@@ -105,4 +106,5 @@ Future<JsObject> embedJsScriptObject(String uri, String jsObjectName,
 
     return ret.future;
 }
+