+++ CHANGELOG.md
@@ -4,12 +4,15 @@
 
 **::Overview::**
 
-Minor bug fixes.
+Added Vimeo support.
 
 **::Details::**
 
+* Vimeo support added.
+* Added cleaner support for handling video players which are Flash objects.
+  Vimeo is using it, still need to migrate YouTube to it.
 * Bug fixes:
-    * (none yet)
+    * More fixes on how YouTube videos clean themselves up when destroyed.
 * Minor cleanup of the CHANGELOG.md.
 
 ## 0.2.0